- convert JoinOptimizer to Visitor
- remove ISqlExpressionWalkable
