+++ Source/LinqToDB/SqlQuery/SqlBinaryExpression.cs
@@ -64,18 +64,6 @@ namespace LinqToDB.SqlQuery
 
 		#endregion
 
-		#region ISqlExpressionWalkable Members
-
-		ISqlExpression ISqlExpressionWalkable.Walk<TContext>(WalkOptions options, TContext context, Func<TContext, ISqlExpression, ISqlExpression> func)
-		{
-			Expr1 = Expr1.Walk(options, context, func)!;
-			Expr2 = Expr2.Walk(options, context, func)!;
-
-			return func(context, this);
-		}
-
-		#endregion
-
 		#region IEquatable<ISqlExpression> Members
 
 		bool IEquatable<ISqlExpression>.Equals(ISqlExpression? other)
