+++ Source/LinqToDB/SqlQuery/SqlCreateTableStatement.cs
@@ -40,12 +40,6 @@ namespace LinqToDB.SqlQuery
 			return writer;
 		}
 
-		public override ISqlExpression? Walk<TContext>(WalkOptions options, TContext context, Func<TContext, ISqlExpression, ISqlExpression> func)
-		{
-			Table = (SqlTable)((ISqlExpressionWalkable)Table).Walk(options, context, func)!;
-			return base.Walk(options, context, func);
-		}
-
 		public override ISqlTableSource? GetTableSource(ISqlTableSource table, out bool noAlias)
 		{
 			noAlias = false;
