+++ Source/LinqToDB/SqlQuery/SqlDropTableStatement.cs
@@ -31,12 +31,6 @@ namespace LinqToDB.SqlQuery
 			return writer;
 		}
 
-		public override ISqlExpression? Walk<TContext>(WalkOptions options, TContext context, Func<TContext, ISqlExpression, ISqlExpression> func)
-		{
-			((ISqlExpressionWalkable?)Table)?.Walk(options, context, func);
-			return base.Walk(options, context, func);
-		}
-
 		public override ISqlTableSource? GetTableSource(ISqlTableSource table, out bool noAlias)
 		{
 			noAlias = false;
