+++ Source/LinqToDB/SqlQuery/SqlInsertStatement.cs
@@ -44,17 +44,6 @@ namespace LinqToDB.SqlQuery
 				.AppendElement(Output);
 		}
 
-		public override ISqlExpression? Walk<TContext>(WalkOptions options, TContext context, Func<TContext, ISqlExpression, ISqlExpression> func)
-		{
-			With?.Walk(options, context, func);
-			((ISqlExpressionWalkable?)_insert)?.Walk(options, context, func);
-			((ISqlExpressionWalkable?)Output)?.Walk(options, context, func);
-
-			SelectQuery = (SelectQuery)SelectQuery.Walk(options, context, func);
-
-			return base.Walk(options, context, func);
-		}
-
 		public override ISqlTableSource? GetTableSource(ISqlTableSource table, out bool noAlias)
 		{
 			noAlias = false;
