+++ Source/LinqToDB/SqlQuery/SqlOrderByItem.cs
@@ -13,11 +13,6 @@ namespace LinqToDB.SqlQuery
 		public ISqlExpression Expression   { get; internal set; }
 		public bool           IsDescending { get; }
 
-		internal void Walk<TContext>(WalkOptions options, TContext context, Func<TContext, ISqlExpression, ISqlExpression> func)
-		{
-			Expression = Expression.Walk(options, context, func)!;
-		}
-
 		#region Overrides
 
 #if OVERRIDETOSTRING
