+++ Source/LinqToDB/SqlQuery/SqlTable.cs
@@ -331,23 +331,6 @@ namespace LinqToDB.SqlQuery
 
 		#endregion
 
-		#region ISqlExpressionWalkable Members
-
-		public virtual ISqlExpression Walk<TContext>(WalkOptions options, TContext context, Func<TContext, ISqlExpression, ISqlExpression> func)
-		{
-			if (TableArguments != null)
-				for (var i = 0; i < TableArguments.Length; i++)
-					TableArguments[i] = TableArguments[i].Walk(options, context, func)!;
-
-			if (SqlQueryExtensions != null)
-				foreach (var e in SqlQueryExtensions)
-					e.Walk(options, context, func);
-
-			return func(context, this);
-		}
-
-		#endregion
-
 		#region System tables
 
 		internal static SqlTable Inserted(EntityDescriptor entityDescriptor)
