+++ Source/LinqToDB/SqlQuery/SqlUpdateStatement.cs
@@ -38,17 +38,6 @@ namespace LinqToDB.SqlQuery
 			return writer;
 		}
 
-		public override ISqlExpression? Walk<TContext>(WalkOptions options, TContext context, Func<TContext, ISqlExpression, ISqlExpression> func)
-		{
-			With?.Walk(options, context, func);
-			((ISqlExpressionWalkable?)_update)?.Walk(options, context, func);
-			((ISqlExpressionWalkable?)Output)?.Walk(options, context, func);
-
-			SelectQuery = (SelectQuery)SelectQuery.Walk(options, context, func);
-
-			return base.Walk(options, context, func);
-		}
-
 		public override ISqlTableSource? GetTableSource(ISqlTableSource table, out bool noAlias)
 		{
 			var result = SelectQuery.GetTableSource(table);
